基本功能完成 停用词过滤没弄好 注释没弄好 面向对象式写法没弄好 而且我觉得PageRank有点问题

+++ src/test/java/CaculatePRtest.java
@@ -22,7 +22,7 @@ public class CaculatePRtest {
          * D  0  1  1  0
          * */
 
-        int[][] link = {
+        double[][] link = {
                 { 0, 1, 1, 0 },
                 { 0, 0, 1, 0 },
                 { 1, 1, 0, 0 },
